+++ src/mc_control/mc_mrqp_controller.cpp
@@ -72,7 +72,7 @@ MCMRQPController::MCMRQPController(const std::vector<std::shared_ptr<mc_rbdyn::R
     mc_solver::Collision("LARM_LINK5", "CHEST_LINK1", 0.05, 0.01, 0.)
   });
 
-  hrp2postureTask = std::shared_ptr<tasks::qp::PostureTask>(new tasks::qp::PostureTask(mrqpsolver->robots.mbs(), hrp2_drc_index, mrqpsolver->robots.robot().mbc().q, 1, 5));
+  hrp2postureTask = std::shared_ptr<tasks::qp::PostureTask>(new tasks::qp::PostureTask(mrqpsolver->robots.mbs(), static_cast<int>(hrp2_drc_index), mrqpsolver->robots.robot().mbc().q, 1, 5));
   std::cout << "MCController(base) ready" << std::endl;
 }
